+++ src/args.c
@@ -8,7 +8,7 @@
 #include <stdarg.h>
 #include <stdint.h>
 #include <assert.h>
-#include "args.h"
+#include <args.h>
 
 
 /* ------------------ */
